+++ CHANGELOG.md
@@ -26,6 +26,7 @@
 * JavaScript client backwards incompatible changes
   * Socket params are now passed to `socket.connect()` instead of an option on the constructor.
   * Socket params are no longer merged as default params for channel params. Use `connect/2` on the server to wire up default channel assigns.
+  * Socket `chan` has been renamed to `channel`, for example `socket.channel("some:topic")`
 
 ## v0.15.0 (2015-7-27)